:recycle: refactor : auth interface 를 type.ts 로 이동
