--- backend/src/auth/auth.interface.ts
@@ -1,8 +1,0 @@
-export interface RefreshTokenWrapper {
-  token: string;
-  isRevoked: boolean;
-}
-
-export interface JwtPayload {
-  userId: string;
-}
