+++ backend/src/util/type.ts
@@ -70,3 +70,12 @@ export interface RelationshipRequest extends VerifiedRequest {
   relationship: Relationship | null;
   targetId: UserId;
 }
+
+export interface RefreshTokenWrapper {
+  token: string;
+  isRevoked: boolean;
+}
+
+export interface JwtPayload {
+  userId: string;
+}
